[change] - bookCar() method - done

+++ backend-service/src/main/java/com/sda/rentalloapp/domain/CarBooking.java
@@ -31,7 +31,7 @@ public class CarBooking {
 
     LocalDate endDate;
 
-    int totalPriceInPolishGrosz;
+    long totalPriceInPolishGrosz;
 
     @ManyToOne
     Address startLocation;
